refactor: improve snaps
